Finished base container for Event Class
